# swagger.yaml
swagger: "2.0"
info:
  description: "This is a phoenix storage service."
  version: "1.0.0"
  title: "phoenix"
  termsOfService: "http://swagger.io/terms/"
  contact:
    name: iotex-dev Team
    email: iotex-dev@iotex.io
    url: https://iotex.io/
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "pg.iotex.io"
basePath: "/v1"
tags:
- name: "pod"
  description: "Everything about pods"
- name: "pea"
  description: "Everything about Pea"
schemes:
- "https"
- "http"
paths:
  /pods:
    post:
      tags:
      - "pod"
      summary: "Add a new pod to the storage"
      description: ""
      operationId: "addPod"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Pod object that needs to be added to the storage"
        required: true
        schema:
          $ref: "#/definitions/Pod"
      responses:
        "200":
          description: "Successful operation"
        "400":
          description: "Invalid name supplied"
        "401":
          description: "Unauthorized Error"  
        "403":
          description: "Permission denied"                   
      security:
      - JWT: []
      
  /pods/{name}:
    delete:
      tags:
      - "pod"
      summary: "Deletes a Pod"
      description: ""
      operationId: "deletePod"
      produces:
      - "application/json"
      parameters:
      - name: "name"
        in: "path"
        description: "Pod name to delete"
        required: true
        type: "string"
      responses:
        "400":
          description: "Invalid name supplied"
        "200":
          description: "Successful operation"
        "401":
          description: "Unauthorized Error"    
        "403":
          description: "Permission denied"                 
      security:
      - JWT: []
  /pea/{podName}/{peaPath}:
    put:
      tags:
      - "pea"
      summary: "uploads an pea object"
      description: ""
      operationId: "uploadFile"
      consumes:
      - "multipart/form-data"
      produces:
      - "application/json"
      parameters:
      - name: "podName"
        in: "path"
        description: "Name of Pod to upload"
        required: true
        type: "string"
      - name: "peaPath"
        in: "path"
        description: "Name of Pea to upload"
        required: true
        type: "string"
      - name: ""
        in: "formData"
        description: "object data to upload"
        required: false
        type: "string"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
        "400":
          description: "Invalid upload supplied"            
        "401":
          description: "Unauthorized Error"     
        "403":
          description: "Permission denied"                
      security:
      - JWT: []
    get:
      tags:
      - "pea"
      summary: "get an pea object"
      description: "get pea object in storage"
      parameters:
      - name: "podName"
        in: "path"
        description: "Name of Pod to upload"
        required: true
        type: "string"
      - name: "peaPath"
        in: "path"
        description: "Name of Pea to upload"
        required: true
        type: "string"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
        "400":
          description: "Invalid get supplied"            
        "401":
          description: "Unauthorized Error"       
        "403":
          description: "Permission denied"              
      security:
      - JWT: []
    delete:
      tags:
      - "pea"
      summary: "delete an pea object"
      description: "delete pea object in storage"
      parameters:
      - name: "podName"
        in: "path"
        description: "Name of Pod to upload"
        required: true
        type: "string"
      - name: "peaPath"
        in: "path"
        description: "Name of Pea to upload"
        required: true
        type: "string"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
        "400":
          description: "Invalid get supplied"
        "401":
          description: "Unauthorized Error"    
        "403":
          description: "Permission denied"                 
      security:
      - JWT: []
  /pea/{podName}:
    get:
      tags:
      - "pea"
      summary: "get pea object list"
      description: "get pea object list with Pod name in storage"
      parameters:
      - name: "podName"
        in: "path"
        description: "Name of Pod to upload"
        required: true
        type: "string"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
        "400":
          description: "Invalid ID supplied"
        "401":
          description: "Unauthorized Error"
        "403":
          description: "Permission denied"    
      security:
      - JWT: []
securityDefinitions:
  JWT:  
    type: apiKey
    name: Authorization
    in: header
        # scopes:
        #   "create:pods": create pods
        #   "delete:pods": delete pods
        #   "read:pods": read pods
        #   "upload:pea": upload pea
        #   "delete:pea": delete pea
        #   "read:pea": read pea
    
definitions:
  Pod:
    type: "object"
    required:
    - "name"
    properties:
      name:
        type: "string"
        example: "doggie"
  ApiResponse:
    type: "object"
    properties:
      name:
        type: "string"
      path:
        type: "string"
      message:
        type: "string"
externalDocs:
  description: "Find out more about Swagger"
  url: "http://swagger.io"